修改

+++ console/src/main/java/org/example/food_demo/console/domain/FoodListVO.java
@@ -6,9 +6,9 @@ import java.util.List;
 @Data
 public class FoodListVO {
     private List<FoodItemVo> list;
-    private Integer total;
+    private Long total;
     private Integer pageSize;
-    public FoodListVO(List<FoodItemVo> list, Integer total, Integer pageSize) {
+    public FoodListVO(List<FoodItemVo> list, Long total, Integer pageSize) {
         this.list = list;
         this.total = total;
         this.pageSize = pageSize;
